For Debian rootfs make sure the default user can sudo
By default, Debian does not install sudo and creates a root user, but since we
do not set any root password, we need a way to access priviliged operations. We
install the sudo package and, since we already add the default user to the sudo
group and the default /etc/sudoers file will allow sudo for all users in the
sudo group, that's all we need.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>