Add firstboot service
To allow generic images for Pine64 and Pinebook, some actions can only be
done when the image is actually booted on real hardware. This commit adds
a simple systemd-powered service with a flag file marker which enables
such actions.

The firstboot condition is true when the flag file at
/etc/pine64-firstboot.d/.firstboot-done does not exist.

The service will run all lsbsysinit compatible executable scripts in
/etc/pine64-firstboot.d in sorted order.

When all scripts have run without error, firstboot is considered
successfull and the marker file is created.