allow-hotplug instead of auto for eth0
This is not the appropriate setting for the pinebook, as it does not have a eth0 connection! It causes an error in the networking.service log instead of clean exit. It will additionally delay boot on pine64s/sopines without a ethernet cable plugged in. 
This is the expected output (http://sprunge.us/NVjZ) as opposed to the output met (http://sprunge.us/EHFW)