Added standard bashrc and mounted it in HOME
The original `make shell` script caused host's bashrc file to be mounted
together with HOME directory and executed when you entered shell.
This has an undesireable side effect in case somoene has their own tools
added to host's bashrc that are nowhere to be found in the docker.

In my case it was liquidprompt and gpg2 which were missing - my
configuration of .bashrc actually caused the initial `make sync` to fail

Having a "standard" bashrc instead makes for consistent experience for
all developers.