Added instructions to enable the onboard IR
The only thing I don't understand is how to use a `&gpio2` label instead of the phandle value. Any attempts to build a new overlay using this syntax results in the following error: `gpio-ir-recv/dtbo: ERROR (phandle_references): Reference to non-existent node or label "gpio2"`. And an addition of `-@` parameter for `dtc` doesn't help either.

Please, help with this question and I'll update pull request if needed

Closes #225